feat(cli)!: promote stats to a top-level command with OGC --dates
`snowtool query stats` becomes `snowtool stats DATASET TRIPLET`, taking
--dates/--years (OGC datetime interval, or MM-DD + --years for day-of-year)
instead of --start/--end/--doy/--start-year/--end-year, and renders a status
spinner on stderr while the query runs. `query` keeps `dates` for now (moves
in Task 8); `_build_query` and the old stats command are deleted from
query.py.
